added popper gem to Gemfile. Changed postgresql port back to default 5432

+++ home_gig_4/db/schema.rb
@@ -18,7 +18,7 @@ ActiveRecord::Schema.define(version: 2019_11_19_163455) do
   create_table "bids", force: :cascade do |t|
     t.string "description"
     t.float "amount"
-    t.bigint "job_id", null: false
+    t.integer "job_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "user_id"
@@ -50,7 +50,7 @@ ActiveRecord::Schema.define(version: 2019_11_19_163455) do
   create_table "reviews", force: :cascade do |t|
     t.string "text"
     t.integer "rating"
-    t.bigint "job_id", null: false
+    t.integer "job_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "user_id"
